UICAL-75 correctly define settings pset
The pset `settings.calendar.enabled` was listed as a subpermission but
never actually defined (it was removed in #194). This caused trouble
when displaying a user with all permissions because the sort function
would explode when it came across the pset with no attributes and it
tried to compare non-existent attributes.

Fixes UICAL-75