UICAL-148 use react-intl for date formatting
Some date display was pretty convoluted, storing formatters as
translations and passing those to moment and passing that to
`dateformat` and ... yeesh. It's like this: use `moment` internally to
store date values; use `react-intl` externally to present values to the
UI, i.e. don't call `moment.format()` for display. This resolves date
strings like `01/20/AMAMAMAM` which passed the `dateFormat` translation
value as an argument to `moment.format()`.

Additionally, use the localized value for `Closed`. It was always
available, but we just weren't using it.

Refs UICAL-148