separate date from display-date
There is not great separation between display-date format, e.g. M/D/YYYY
for the locale `en-US` and the internal formatting of dates, e.g.
`MM/DD/YYYY` in the locale `en-US`. The adjustments here attempt to sync
that up, looking for the former when dates are displayed as text but
using the latter when they are input as values in forms.