UICAL-227 test SortableMultiColumnList
Test `./src/components/SortableMultiColumnList`. Export some functions
that were previously private but external to the component itself. I
don't love exporting these functions and _not_ actually testing the
component itself, but this'll just become a really complicated test of
the internals of MCL if we try to test the component. Literally all this
component does is reshape its data and pass it through to MCL so I don't
actually feel all that guilty. Well, a little guilty. :shrug:

Refs UICAL-227