restore @types/* to direct deps (#440)
Restore direct `@types` deps that were moved to dev-deps in #439. This module is not transpiled upon release; rather, it is transpiled when building a bundle. Hence, the dependencies relevant to transpilation must be available to the bundler (i.e. must be available when platform-level dependencies are installed) and thus must be listed here as direct deps.